// src/features/employees/models/employee.ts
export default interface Employee {
  id: string;
  employeeId: string;
  firstName: string;
  middleName: string;
  lastName: string;
  gender: string;
  email: string;
  phone: string;
  avatar: string;
  permanentAddress: string;
  currentAddress: string;
  dateOfJoining: Date;
  group: string;
  subGroup: string;
  designation: string;
  createdAt: Date;
  updatedAt: Date;
}